reemplaza conjunto de awaits por promise.all, reubica metodo parseTelefono
